Move `FeedImageDataLoaderWithFallBackComposite` to production
