+++ EssentialApp/EssentialApp/FeedImageDataLoaderWithFallBackComposite.swift
@@ -1,0 +1,40 @@
+//
+//  FeedImageDataLoaderWithFallBackComposite.swift
+//  EssentialAppTests
+//
+//  Created by Jair Moreno Gaspar on 08/02/21.
+//
+
+import Foundation
+import EssentialFeed
+
+public class FeedImageDataLoaderWithFallBackComposite: FeedImageDataLoader {
+    private let primaryLoader: FeedImageDataLoader
+    private let fallbackLoader: FeedImageDataLoader
+    
+    public init(primaryLoader: FeedImageDataLoader, fallbackLoader: FeedImageDataLoader) {
+        self.primaryLoader = primaryLoader
+        self.fallbackLoader = fallbackLoader
+    }
+    
+    private class TaskWrapper: FeedImageDataLoaderTask {
+        var wrapped: FeedImageDataLoaderTask?
+        
+        func cancel() {
+            wrapped?.cancel()
+        }
+    }
+    
+    public func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        let task = TaskWrapper()
+        task.wrapped = primaryLoader.loadImageData(from: url) { [weak self] result in
+            switch result {
+            case .failure:
+                task.wrapped = self?.fallbackLoader.loadImageData(from: url, completion: completion)
+            case .success:
+                completion(result)
+            }
+        }
+        return task
+    }
+}
